Move other than app.module.ts modules out of the root folder.
